requirements.txt
asgiref==3.8.1
Django==5.1.1
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
packaging==24.1
sqlparse==0.5.1
tzdata==2024.1
